+++ src/packer.h
@@ -93,10 +93,10 @@ class cd::Packer
 
 
 ///@brief Add data to pack in packer data structure.
-    virtual uint32_t Add(uint32_t id, uint32_t length, const void *position);
+    virtual uint32_t Add(uint32_t id, uint64_t length, const void *position);
 
 ///@brief Get total size required for table (metadata) and data.
-    virtual char *GetTotalData(uint32_t &total_data_size);
+    virtual char *GetTotalData(uint64_t &total_data_size);
 
 ///@brief Grow buffer size internally used in packer.
     virtual void SetBufferGrow( uint32_t table_grow_unit, uint32_t data_grow_unit);
